Update libtiff to 4.3.0
Changes to libtiff for gpdl.

1) Ensure that libtiff doesn't mess with 'boolean' in GS builds
on Windows. Without this, the jpeg structures used by our JPEG
lib build are different in size when called from gs and libtiff,
resulting in runtime errors.

2) Update libtiff so that it can correctly call into the jpeg
library so that memory operations happen from our pools, not
malloc/free. Slightly horrid in that this is more complex with
OJPEG than JPEG files.

Fix some compiler warnings stemming from libtiff update

Various tweaks for libtiff 4.3.0 to build on Windows

Mainly "importing" tif_config.vc.h and tiffconf.vc.h - the tiff build now
created those with cmake.

Then typedef'ing several C99 ***_t style types.

Finally defining matching C99 PRI* printf formatting macros

+++ tiff/test/ascii_tag.c
@@ -108,8 +108,8 @@ main()
 	for (i = 0; i < NTAGS; i++) {
 		if (!TIFFSetField(tif, ascii_tags[i].tag,
 				  ascii_tags[i].value)) {
-			fprintf(stderr, "Can't set tag %lu.\n",
-				(unsigned long)ascii_tags[i].tag);
+			fprintf(stderr, "Can't set tag %"PRIu32".\n",
+				ascii_tags[i].tag);
 			goto failure;
 		}
 	}
@@ -144,8 +144,8 @@ main()
 	for (i = 0; i < NTAGS; i++) {
 		if (!TIFFGetField(tif, ascii_tags[i].tag, &value)
 		    || strcmp(value, ascii_tags[i].value)) {
-			fprintf(stderr, "Can't get tag %lu.\n",
-				(unsigned long)ascii_tags[i].tag);
+			fprintf(stderr, "Can't get tag %"PRIu32".\n",
+				ascii_tags[i].tag);
 			goto failure;
 		}
 	}
